+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-DAo5Lt0P.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-PEkgECve.js";import{_ as a}from"./mermaid.core-Bd0teJvs.js";import"./chunk-5VM5RSS4-BC-m68-j.js";import"./chunk-XXDRQBXY-BXRn8_vu.js";import"./chunk-POPQ4Y6H-chNZGPYR.js";import"./chunk-F27PBJKO-DVVabRS4.js";import"./channel-ByuDC_tH.js";import"./index-CXqveIBt.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};